refactor: split content index parsing
